complete base dokeos addUser
git-svn-id: http://192.168.1.235/svn/repo/octopus@3273 7df5b637-c7b4-2643-9bff-991d5fa3ecdc

+++ src/com/itdaoshi/discuz/dao/CdbUcMembersDAObject.java
@@ -51,7 +51,7 @@ public class CdbUcMembersDAObject extends DiscuzDAO {
     CdbUcMembers obj = (CdbUcMembers) obj1;
     synchronized (conn) {
       try {
-         Integer nextID = getNextPrimaryID();
+         Long nextID = getNextPrimaryID();
 //        Integer nextID = 2;
         sqlStat.append("INSERT ");
         sqlStat
@@ -88,7 +88,7 @@ public class CdbUcMembersDAObject extends DiscuzDAO {
   }
 
   @Override
-  protected Integer getNextPrimaryID()   {
+  protected Long getNextPrimaryID()   {
     QueryRunner run = new QueryRunner();
     ResultSetHandler h = new ResultSetHandler() {
       public Object handle(ResultSet rs) throws SQLException {
@@ -109,7 +109,7 @@ public class CdbUcMembersDAObject extends DiscuzDAO {
   };
     try{
       Object[] result = (Object[]) run.query( conn, "SELECT MAX(uid) FROM CDB_UC_MEMBERS ", h);
-      return (Integer)result[0]+1;
+      return (Long)result[0]+1;
     // do something with the result
     }catch (Exception e) {
      e.printStackTrace();
